+++ src/main/java/com/heroiclabs/nakama/Channel.java
@@ -33,12 +33,34 @@ public class Channel {
      * The server-assigned channel ID.
      */
     private String id;
+
     /**
      * The presences visible on the chat channel.
      */
     private List<UserPresence> presences;
+
     /**
      * The presence of the current user. i.e. Your self.
      */
     private UserPresence self;
+
+    /**
+     * The name of the chat room, or an empty string if this message was not sent through a chat room.
+     */
+    private String roomName;
+
+    /**
+     *  The ID of the group, or an empty string if this message was not sent through a group channel.
+     */
+    private String groupId;
+
+    /**
+     * The ID of the first DM user, or an empty string if this message was not sent through a DM chat.
+     */
+    private String userIdOne;
+
+    /**
+     * The ID of the second DM user, or an empty string if this message was not sent through a DM chat.
+     */
+    private String userIdTwo;
 }