+++ src/main/java/com/heroiclabs/nakama/MatchData.java
@@ -46,7 +46,7 @@ public class MatchData {
     /**
      * Information on the user who sent the state change.
      */
-    private UserPresence userPresence;
+    private UserPresence presence;
 
     /**
      * Returns match data
@@ -58,4 +58,9 @@ public class MatchData {
         }
         return BaseEncoding.base64().decode(this.data);
     }
+
+    /**
+     * True if this data was delivered reliably, false otherwise.
+     */
+    private boolean reliable;
 }